neuracore_new_data_format/pyproject.toml
[project]
name = "neuracore-new-data-format"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "h5py>=3.14.0",
    "mcap>=1.3.0",
    "pydantic>=2.11.9",
    "av==15.1.0",
    "protobuf>=5.27.2",
    "numpy>=2.0.1",
    "mcap-protobuf-support>=0.5.3",
    "mypy-protobuf>=3.6.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "grpcio-tools>=1.65.1",
]

[build-system]
requires = ["uv_build>=0.8.20,<0.9.0"]
build-backend = "uv_build"
